+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0-rc.2]
 
+### Fixed
+
+- Async validation works again for custom validators and `superValidate`.
+
 ### Added
 
-- `reset` now have an additional `data` option that can be used to re-populate the form with data.
+- `reset` now has an additional `data` option that can be used to re-populate the form with data.
 - `intProxy`, `numberProxy`, `dateProxy` and `stringProxy` now have an `empty` option, so empty values can be set to `null` or `undefined`.
 
 ## [1.0.0-rc.1]
@@ -19,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Explicitly setting a form `id` for multiple forms is not required anymore when using `use:enhance`, unless the forms are using the same schema. An id can be specified in the options or in a hidden form field called `__superform_id`.
 - `setError` doesn't handle form-level errors anymore, use refine/superRefine on the schema, or the `message` helper.
 - `FieldPath` is gone - the following methods are now using a string accessor like `tags[2].id` instead of an array like `['tags', 2, 'id']`: `validate`, `setError` and all proxy methods (ending with `Proxy`). This also applies to generic components.
-- The signature for `allErrors` and `firstError` has changed to `{ path: string[]; messages: string[] }`.
+- The signature for `allErrors` and `firstError` has changed to `{ path: string; messages: string[] }`.
 - The literal `"any"` is now an allowed value in `step` for constraints.
 - Multiple `regex` and `step` is now allowed in a schema. A warning will be emitted by default, that can be turned off.
 - The signature for `options.resetForm` has changed to `boolean | () => boolean` (it was async before).